PowerVR Graphics SDK v3.5 Release.
The 3.5 SDK is distributed under the MIT license

+++ Examples/Advanced/Water/OGLES2/OGLES2Water.cpp
@@ -1920,7 +1920,7 @@ void OGLES2Water::DrawMesh(int i32NodeIndex, const ModelShader& shaderProgram)
 			if(m_apuiModelIndexVbo[i32MeshIndex])
 			{
 				// Indexed Triangle strips
-				glDrawElements(GL_TRIANGLE_STRIP, pMesh->pnStripLength[i]+2, GL_UNSIGNED_SHORT, (GLshort*)(offset*2));
+				glDrawElements(GL_TRIANGLE_STRIP, pMesh->pnStripLength[i] + 2, GL_UNSIGNED_SHORT, (GLshort*)(size_t)(offset * 2));
 			}
 			else
 			{
