+++ Tools/PVRTModelPOD.cpp
@@ -3412,7 +3412,7 @@ static void InterleaveArray(
 		memcpy(pInterleaved + i * nStride + nOffset, (char*)data.pData + i * data.nStride, data.nStride);
 
 	FREE(data.pData);
-	data.pData		= (unsigned char*)nOffset;
+	data.pData		= (unsigned char*)(size_t)nOffset;
 	data.nStride	= nStride;
 	nOffset			+= PVRTModelPODDataStride(data) + nPadding;
 }
